api improvements to make it more clear that the program creator functions are NOT imperative. the inspiration for this change is thanks to similar efforts by code.world. Code.world named their functions drawingOf, animationOf, simulationOf, interactionOf. I decided to suffix with Program to map to the convention already set in place by elms standard libraries.
Additionally, I renamed the module to be Collage.Program instead of Colage.Interaction. The library really is all about creating graphics programs, so i think that name just makes a little bit more sense.

Finally, after changing the name of the module, I finally felt like I had a reason to remove the overlay function. I still think that the overlay function is helpful and necessary for the course, but I have decided to move it to an elm-graphics-extra module.